fixed getEmployeById function and added test to deleteEmploye function

+++ src/main/java/tn/esprit/spring/services/EmployeServiceImpl.java
@@ -72,6 +72,8 @@ public class EmployeServiceImpl implements IEmployeService {
 	@Transactional	
 	public void affecterEmployeADepartement(int employeId, int depId) {
 		try {
+			if(deptRepoistory.findById(depId).isPresent() && employeRepository.findById(employeId).isPresent())
+			{	
 			Departement depManagedEntity = deptRepoistory.findById(depId).get();
 			Employe employeManagedEntity = employeRepository.findById(employeId).get();
 			if(depManagedEntity.getEmployes() == null){
@@ -85,6 +87,7 @@ public class EmployeServiceImpl implements IEmployeService {
 				l.debug("Employe "+ employeManagedEntity.getNom() +"has been affected to the department "
 						+ depManagedEntity.getName());
 			}
+			}
 		} catch (Exception e) {
 			l.error(e.getMessage());
 		}
@@ -96,14 +99,21 @@ public class EmployeServiceImpl implements IEmployeService {
 	public void desaffecterEmployeDuDepartement(int employeId, int depId)
 	{
 		try {
-			Departement dep = deptRepoistory.findById(depId).get();
-			int employeNb = dep.getEmployes().size();
-			for(int index = 0; index < employeNb; index++){
-				if(dep.getEmployes().get(index).getId() == employeId){
-					l.debug("Employe" + dep.getEmployes().get(index).getNom() + "kicked from the department " + dep.getName() );
-					dep.getEmployes().remove(index);
-					break;
+			if(deptRepoistory.findById(depId).isPresent())
+			{
+				Departement dep = deptRepoistory.findById(depId).get();
+				int employeNb = dep.getEmployes().size();
+				for(int index = 0; index < employeNb; index++){
+					if(dep.getEmployes().get(index).getId() == employeId){
+						l.debug("Employe" + dep.getEmployes().get(index).getNom() + "kicked from the department " + dep.getName() );
+						dep.getEmployes().remove(index);
+						break;
+					}
 				}
+			}
+			else
+			{
+				l.error("the departement does not exist");
 			}
 		} catch (Exception e) {
 			l.error(e.getMessage());
@@ -114,6 +124,7 @@ public class EmployeServiceImpl implements IEmployeService {
 	public Contrat getContratByReference(int reference) {
 		return contratRepoistory.findById(reference).get();
 	}
+	
 	public int ajouterContrat(Contrat contrat) {
 		try{
 			if(contratRepoistory.save(contrat) != null)
@@ -130,7 +141,6 @@ public class EmployeServiceImpl implements IEmployeService {
 		try {
 			Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
 			Employe employeManagedEntity = employeRepository.findById(employeId).get();
-
 			contratManagedEntity.setEmploye(employeManagedEntity);
 			if(contratRepoistory.save(contratManagedEntity) != null){
 				l.debug("Contract " + contratManagedEntity.getReference() + " affected to employe "+ employeManagedEntity.getNom());
@@ -150,18 +160,26 @@ public class EmployeServiceImpl implements IEmployeService {
 		}
 		return testValidator;
 	}
+	
 	public void deleteEmployeById(int employeId)
 	{
 		try{
-			Employe employe = employeRepository.findById(employeId).get();
-			//Desaffecter l'employe de tous les departements
-			//c'est le bout master qui permet de mettre a jour
-			//la table d'association
-			for(Departement dep : employe.getDepartements()){
-				dep.getEmployes().remove(employe);
-				l.debug("Employe "+ employe.getNom() + " from the department "+ dep.getName());
+			if(employeRepository.findById(employeId).isPresent())
+			{
+				Employe employe = employeRepository.findById(employeId).get();
+				//Desaffecter l'employe de tous les departements
+				//c'est le bout master qui permet de mettre a jour
+				//la table d'association
+				for(Departement dep : employe.getDepartements()){
+					dep.getEmployes().remove(employe);
+					l.debug("Employe "+ employe.getNom() + " from the department "+ dep.getName());
+				}
+				employeRepository.delete(employe);
 			}
-			employeRepository.delete(employe);
+			else
+			{
+				l.error("no such element");
+			}
 		}catch (Exception e) {
 			l.error(e.getMessage());
 		}
@@ -218,7 +236,15 @@ public class EmployeServiceImpl implements IEmployeService {
 
 	@Override
 	public Employe getEmployeById(int id) {
-		return employeRepository.findById(id).get();
+		if(employeRepository.findById(id).isPresent())
+		{
+			return employeRepository.findById(id).get();
+		}
+		else
+		{
+			return null;
+		}
+		
 	}
 
 }
